[#41] css 코드 분리
