+++ src/pages/chat/[id].tsx
@@ -1,6 +1,6 @@
 import React from 'react';
+import ChatDetail from 'src/components/Chat/Chat';
 import Layout from 'src/components/shared/Layout';
-import ChatDetail from 'src/components/chat/Chat';
 
 const chatDetail = () => (
   <Layout noMenu bgColor back title="채팅">
