+++ src/pages/chat/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
+import ChatList from 'src/components/Chat/ChatList';
 import Layout from 'src/components/shared/Layout';
-import ChatList from 'src/components/chat/ChatList';
 
 const chat = () => (
   <Layout noMenu bgColor back title="채팅">
